% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anti_data}
\alias{anti_data}
\title{BGC list from antiSMASH for S.coelicolor}
\format{
A data frame with 27 rows and 6 variables:
\describe{
  \item{Cluster}{Cluster #, numeric}
  \item{Start}{Start location of a cluster, in bp}
  \item{Stop}{Stop location of a cluster, in bp}
  \item{Type}{Type of a cluster}
  \item{chromosome}{chromosome, on which cluster is, filled automatically ipon file upload}
  \item{Type2}{Type of a cluster, filled automatically upon file upload. Same as Type}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_antismash.csv}
}
\usage{
anti_data
}
\description{
Dataset containing list of clusters from antiSMASH5, found in S.coelicolor
}
\keyword{datasets}
