% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_transformation.R
\name{antismash_to_csv}
\alias{antismash_to_csv}
\title{antismash_to_csv}
\usage{
antismash_to_csv(file, write_to = getwd())
}
\arguments{
\item{file}{- path to a json file,}

\item{write_to}{- path where to write generated csv file}
}
\value{
csv file in specified location
}
\description{
Function, that returns dataframe, out of supplied antismash json file
}
