% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arts_data}
\alias{arts_data}
\title{Gene list from ARTS for S.coelicolor}
\format{
A data frame with 161 rows and 13 variables:
\describe{
  \item{Cluster}{Gene #, numeric}
  \item{Start}{Start location of a gene, in bp}
  \item{Stop}{Stop location of a gene, in bp}
  \item{Type}{Type of a gene}
  \item{Hit}{Unique core gene hit identifier, based on count of genes}
  \item{Type2}{Type of a gene, same as Type}
  \item{Core}{Type of core model from ARTS}
  \item{Description}{Annotation of a gene}
  \item{Count}{Count of a gene hit, to keep track of multiple hits}
  \item{ID}{Gene #, same as Cluster}
  \item{Evalue}{E-value of a resistance gene hit}
  \item{Bitscore}{Bitscore of a resistance gene hit}
  \item{Model}{Type of Model of gene hit. If core gene, then the model is Core}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_arts.csv}
}
\usage{
arts_data
}
\description{
Dataset containing list of core and resistance genes from ARTS, found in S.coelicolor. 
It is generated automatically upon zip folder from ARTS upload
}
\keyword{datasets}
