% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_transformation.R
\name{arts_to_csv}
\alias{arts_to_csv}
\title{arts_to_csv}
\usage{
arts_to_csv(project_archive, write_to = getwd())
}
\arguments{
\item{project_archive}{- path to zip file, downloaded from ARTS}

\item{write_to}{- path where to write generated csv file}
}
\value{
csv file in specified location
}
\description{
Function, which extracts tables from arts result zip archive and transforms them into BGCViz input
}
\examples{
\dontrun{
arts_to_csv(<zip-file>)
}

}
