% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_transformation.R
\name{data_to_json}
\alias{data_to_json}
\title{all data to json}
\usage{
data_to_json(csv_file)
}
\arguments{
\item{csv_file}{path to csv}
}
\value{
json
}
\description{
Function, function that takes csv file and converts it to json
}
