% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_data}
\alias{deep_data}
\title{BGC list from DeepBGC for S.coelicolor}
\format{
A data frame with 172 rows and 28 variables:
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_deep.tsv}
}
\usage{
deep_data
}
\description{
Dataset containing list of clusters from DeepBGC, found in S.coelicolor
For full description see DeepBGC documentation!
}
\keyword{datasets}
