% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gecco_data}
\alias{gecco_data}
\title{BGC list from GECCO for S.coelicolor}
\format{
A data frame with 27 rows and 6 variables:
\describe{
  \item{Cluster}{Cluster #, numeric}
  \item{Start}{Start location of a cluster, in bp}
  \item{Stop}{Stop location of a cluster, in bp}
  \item{Type}{Type of a cluster}
  \item{chromosome}{chromosome, on which cluster is, filled automatically ipon file upload}
  \item{Type2}{Type of a cluster, filled automatically upon file upload. Same as Type}
  \item{ID}{Cluster #. Same as Cluster}
  \item{average_p}{Average probability of a cluster}
  \item{max_p}{Maximum probability of a cluster}
  \item{bgc_id}{ID of a cluster, filled by GECCO}
  \item{sequence_id}{ID of annotated sequence}
  \item{type}{Type of a cluster, filled by GECCO}
  \item{proteins}{list of identified proteins in a cluster, filled by GECCO}
  \item{domains}{list of identified domains in a cluster, filled by GECCO}
  \item{pks}{Score that cluster is pks, filled by GECCO}
  \item{other}{Score that cluster is other, filled by GECCO}
  \item{nrps}{Score that cluster is nrps, filled by GECCO}
  \item{alkaloid}{Score that cluster is alkaloid, filled by GECCO}
  \item{terpene}{Score that cluster is terpene, filled by GECCO}
  \item{saccharide}{Score that cluster is sacharide, filled by GECCO}
  \item{ripp}{Score that cluster is ripp, filled by GECCO}
  \item{num_prot}{Number of proteins in a cluster}
  \item{num_domains}{Number of domains in a cluster}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_gecco.tsv}
}
\usage{
gecco_data
}
\description{
Dataset containing list of clusters from GECCO, found in S.coelicolor
}
\keyword{datasets}
