% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_defaults}
\alias{get_defaults}
\title{get_defaults}
\usage{
get_defaults(write_to = getwd())
}
\arguments{
\item{write_to}{- location where the write configuration file. default - current working directory}
}
\value{
csv file in specified location
}
\description{
Function, which downloads default options csv file into provided location.
Then can be used with `set_defaults` function to change the default behaviour
}
