% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_dissect_example}
\alias{get_dissect_example}
\title{get_dissect_example}
\usage{
get_dissect_example(write_to = getwd())
}
\arguments{
\item{write_to}{- path to write csv file to.}
}
\value{
csv file, written to package settings
}
\description{
Function, which downloads a csv file for dissect.py into specified path.
Use to separate regions into separate clusters
}
\examples{
\dontrun{
get_dissect_example(write_to)
}

}
