% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_data}
\alias{prism_data}
\title{BGC list from PRISM for S.coelicolor}
\format{
A data frame with 19 rows and 4 variables:
\describe{
  \item{Cluster}{Cluster #, numeric}
  \item{Start}{Start location of a cluster, in bp}
  \item{Stop}{Stop location of a cluster, in bp}
  \item{Type}{Type of a cluster}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_prism.json}
}
\usage{
prism_data
}
\description{
Dataset containing list of clusters from PRISM, found in S.coelicolor
Generated automatically from json results file
}
\keyword{datasets}
