% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_supp_data}
\alias{prism_supp_data}
\title{Gene list from PRISM for S.coelicolor}
\format{
A data frame with 125 rows and 9 variables:
\describe{
  \item{Cluster}{Gene #, numeric}
  \item{Start}{Start location of a gene, in bp}
  \item{Stop}{Stop location of a gene, in bp}
  \item{Type}{Type of a gene}
  \item{Full_name}{Annotation of a gene hit}
  \item{ID}{Gene #, same as Cluster}
  \item{Type2}{Type of a gene. Same as Type}
  \item{Score}{Score of a gene hite}
  \item{Name}{Name of model, or gene for a hit}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_prism.json}
}
\usage{
prism_supp_data
}
\description{
Dataset containing list of regulatory and resistance from PRISM, found in S.coelicolor.
Generated automatically from json results file
}
\keyword{datasets}
