% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_transformation.R
\name{prism_to_csv}
\alias{prism_to_csv}
\title{prism_to_csv}
\usage{
prism_to_csv(file, write_to = getwd())
}
\arguments{
\item{file}{- path to a json file,}

\item{write_to}{- path where to write generated csv file}
}
\value{
csv file in specified location
}
\description{
Function, that transforms prism json object into dataframe, which could be written to the csv file
}
\examples{
\dontrun{
prism_to_csv(<json-file>)
}
}
