% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reading_functions.R
\name{read_compare}
\alias{read_compare}
\title{#' @description A function, that reads RiPPMiner-Genome file txt
#'
#' @return csv file
#'
#' @noRd}
\usage{
read_compare(data)
}
\description{
#' @description A function, that reads RiPPMiner-Genome file txt
#'
#' @return csv file
#'
#' @noRd
}
