% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rre_data}
\alias{rre_data}
\title{RRE list from RREFinder for S.coelicolor}
\format{
A data frame with 27 rows and 6 variables:
\describe{
  \item{Cluster}{RRE #, numeric}
  \item{Start}{Start location of a RRE, in bp}
  \item{Stop}{Stop location of a RRE, in bp}
  \item{Type}{Type of a RRE}
  \item{chromosome}{chromosome, on which RRE is, filled automatically ipon file upload}
  \item{Type2}{Type of a RRE, filled automatically upon file upload. Same as Type}
  \item{Sequence}{Annotated sequence ID}
  \item{Locus_tag}{Locus tag of a RRE element}
  \item{BGC.ID}{Id, of a BGC}
  \item{BGC.product}{Product of BGC}
  \item{Domain.name}{Domain name of RRE element}
  \item{E.value}{E-value of a hit}
  \item{Bitscore}{Bitscore of a hit}
  \item{End}{End of RRE element, bp}
  \item{ID}{RRE ID, same as Cluster}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_rre.txt}
}
\usage{
rre_data
}
\description{
Dataset containing list of RRE elements from RREFinder, found in S.coelicolor
}
\keyword{datasets}
