% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sempi_data}
\alias{sempi_data}
\title{BGC list from SEMPI for S.coelicolor}
\format{
A data frame with 33 rows and 5 variables:
\describe{
  \item{Cluster}{Cluster #, numeric}
  \item{Start}{Start location of a cluster, in bp}
  \item{Stop}{Stop location of a cluster, in bp}
  \item{Type}{Type of a cluster}
  \item{Type2}{Type of a cluster, filled automatically upon file upload. Same as Type}
}
}
\source{
\url{https://raw.githubusercontent.com/pavlohrab/BGCViz-datasets/main/example_data/sco_sempi.csv}
}
\usage{
sempi_data
}
\description{
Dataset containing list of clusters from SEMPI, found in S.coelicolor
}
\keyword{datasets}
