% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{set_defaults}
\alias{set_defaults}
\title{set_defaults}
\usage{
set_defaults(csv_file)
}
\arguments{
\item{csv_file}{- path to csv file with default options.}
}
\value{
csv file, written to package settings
}
\description{
Function, which uploads default options csv file to the package.
Use with `get_defaults` function to download currently used default options.
}
\examples{
\dontrun{
set_defaults(<csv-file>)
}

}
